scarf: fix -Wnarrowing warnings

+++ scarf/scarf.cc
@@ -338,12 +338,12 @@ a_c_c alm2phase_ginfo(sharp_geom_info *ginfo, const a_c_c &alm, const int64_t lm
       npix += ginfo->nph(i);
     }
 
-    size_t nchunks;
-    size_t chunksize;
+    long unsigned int nchunks;
+    long unsigned int chunksize;
     get_singular_chunk_info(ginfo_new->npairs(), (0==0) ? 128 : 64, 
         nchunks,chunksize);
 
-    auto phase_a2p = get_optional_Pyarr<complex<double>>(out, {2*chunksize, mmax+1, 1});
+    auto phase_a2p = get_optional_Pyarr<complex<double>>(out, {2*chunksize, unsigned(mmax)+1, 1});
     auto phase_mav_a2p = to_mav<complex<double>, 3>(phase_a2p, true);
 
     py::buffer_info buf = phase_a2p.request();
@@ -423,12 +423,12 @@ a_c_c map2phase_ginfo(sharp_geom_info *ginfo, a_d_c &map, size_t lmax, size_t mm
   auto ar = alm.mutable_unchecked<1>();
 
 
-  size_t nchunks;
-  size_t chunksize;
+  long unsigned int nchunks;
+  long unsigned int chunksize;
   get_singular_chunk_info(ginfo_new->npairs(), (0==0) ? 128 : 64, 
       nchunks,chunksize);
 
-  auto phase_m2p = get_optional_Pyarr<complex<double>>(out, {2*chunksize, mmax+1, 1});
+  auto phase_m2p = get_optional_Pyarr<complex<double>>(out, {2*chunksize, unsigned(mmax)+1, 1});
   auto phase_mav = to_mav<complex<double>, 3>(phase_m2p, true);
 
   py::buffer_info buf = phase_m2p.request();
